Move simulation into event-based Simulation class
The ScenarioController and ScenarioView classes are removed in favor of
a Simulation class that emits events for new data and play/pause/reset
and a more general Scenario class that basically handles the view part
at the moment. The controller part is most likely not specific to a
scenario such that it can later be moved into a single controller class
for the whole exhibit.